Improved extension mechanism.
Native functions are now implemented as slots of a global Extensions object.

Ruby code only needs to require loader.rb, which automatically checks if the
native extension library is present, and adds canned versions of the extension
functions if it is missing.

All the functions are inserted into a global $ext object, and some can be added
in more appropriate places.

At the moment, only the exp_blur function is present, which is also exposed as
a member of Qt::Image.